Création des formes dans le model, need affichage pour terminer la base

+++ src/model/Formes/Coord.java
@@ -5,15 +5,15 @@ package ProjetPatron.src.model.Formes;
  */
 public class Coord {
 	
-	private float x;
-	private float y;
+	private int x;
+	private int y;
 
 	/***
 	 * Permet de créer une coordonée (un point)
 	 * @param x la position x
 	 * @param y la position y
 	 */
-	public Coord(float x, float y) {
+	public Coord(int x, int y) {
 		this.x = x;
 		this.y = y;
 	}
@@ -22,7 +22,7 @@ public class Coord {
 	 * Permet de récupérer le x du point
 	 * @return la position x
 	 */
-	public float getX() {
+	public int getX() {
 		return x;
 	}
 
@@ -30,7 +30,7 @@ public class Coord {
 	 * Permet de changer la position x du point
 	 * @param x la nouvelle position x
 	 */
-	public void setX(float x) {
+	public void setX(int x) {
 		this.x = x;
 	}
 
@@ -38,7 +38,7 @@ public class Coord {
 	 * Permet de récupérer le y du point
 	 * @return la position y
 	 */
-	public float getY() {
+	public int getY() {
 		return y;
 	}
 
@@ -46,7 +46,7 @@ public class Coord {
 	 * Permet de changer la position y du point
 	 * @param y la nouvelle position y
 	 */
-	public void setY(float y) {
+	public void setY(int y) {
 		this.y = y;
 	}
